backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
ultralytics
pillow
opencv-python